addressing issues in PR #232 (please read desc.)
- Exp removal handler moved to messageCount.ts, as suggested, with a messageDelete event listener
- SageUser import removed completely from serverLog.ts for the above reason
- Level decreases if current exp exceeds the level exp
- Exp is not detracted if a command message is deleted (standgap until slash commands are added)

NOTE: Looking into the processMessageDelete issue that Josh responded to, this is due to the message being a partial, with ONLY the ID of the message, channel and timestamp sent being available (I will send a screenshot of example partial msg data into the PR thread), all other fields are null. As a standgap measure I've implemented a try-catch into both the serverLog and messageCount function that does the following:
- serverLog: Sends the normal embed but omits author and message content
- messageCount: Does not detract any exp because it is unable to find the author of the message

Given the circumstances of Sage as normal, I doubt this will be a major problem once S22 begins (hence the try-catch implementation), but please let me know with any thoughts regarding this.